fixed bug updating for Calculator Screen, minor changes

+++ app/src/main/kotlin/kaczmarek/moneycalculator/sessions/data/SessionsStorageImpl.kt
@@ -11,7 +11,7 @@ class SessionsStorageImpl(private val sessionsDao: SessionsDao) : SessionsStorag
     }
 
     override suspend fun deleteSessionById(id: SessionId) {
-        sessionsDao.deleteSessionById(id.value)
+        sessionsDao.deleteSessionById(id.value.toInt())
     }
 
     override suspend fun saveSession(
